+++ src/idc/imgstats/class_lister.py
@@ -3,6 +3,9 @@ from typing import List, Dict
 
 def list_classes() -> Dict[str, List[str]]:
     return {
+        "seppl.io.Filter": [
+            "idc.imgstats.filter",
+        ],
         "seppl.io.Writer": [
             "idc.imgstats.writer",
         ],